wow what an exciting test failure
Apparently mod statements can affect how rust picks which name to use in
errors when identifying a specific type. Whenever the `pyo3` feature was
enabled it would pick `ErrReport` instead of `Report`. Fixed it by
moving the `cfg` statement to the end of the file...

Don't write tests that depend on unstable compiler output folks. Don't
be like me.